pgrokd: restructure directory layout
