+++ .gitignore
@@ -1,27 +1,16 @@
-# Binaries for programs and plugins
-*.exe
-*.exe~
-*.dll
-*.so
-*.dylib
+# Artifacts
+.bin
+dist
 
-# Test binary, built with `go test -c`
-*.test
-
-# Output of the go coverage tool, specifically when used with LiteIDE
-/coverage
-*.out
-
-# Application data
+# Configuration and application files
 /pgrokd.yml
 /pgrok.yml
-/data
-
-# Dependency directories (remove the comment below to include it)
-# vendor/
 /.idea
+.envrc
+.task
+
+# System junk
 .DS_Store
 
-.envrc
-.bin
-.task
+# Dependencies
+node_modules
